requirements.txt
Django>=4.2
djangorestframework
django-filter
Pillow

python-dotenv>=1.0.0
python-decouple>=3.6
